Refactored ProjectPage into class component. App refactored to functional component.
