+++ .gitignore
@@ -23,6 +23,3 @@ yarn-debug.log*
 yarn-error.log*
 .idea/
 .python-version
-
-# Not ready to track.
-ProjectPage.js
